ref(spans): Move tests to separate file
